changed data type to store long values

+++ BotsSDK/korebotsdklib/src/main/java/kore/botssdk/models/BotRequest.java
@@ -15,7 +15,7 @@ public class BotRequest extends BaseBotMessage {
     private RestResponse.BotMessage message;
     private String resourceid = "/bot.message";
     private BotInfoModel botInfo;
-    private int id = 1;
+    private long id = 1;
 
     public void setMessage(RestResponse.BotMessage message) {
         this.message = message;
